[OM-62564]: XL AO/Repo: Sort entities by severity in repository.
Intent
======
Make search for entities sorted by severity faster and less resource-intensive.

Background
==========
Today we sort by severity in three steps:
1) Run the search using the repository and return all matching OIDs to the API component.
2) Send the OIDs to the action orchestrator for pagination by severity.
3) Take the next page, and retrieve the entities from the repository.

This is fairly quick, but in a large environment with lots of users it is a bad use of resources, and is unnecessarily complex.

Solution
========
The action orchestrator now broadcasts entity severity updates. This happens at two times:
1) Every time a new action plan is processed.
2) Every time an action completes.

I'm adding an **EntitySeverityClientCache** which can be used by a client to keep entity severities locally accessible. The size of this cache is fairly small - one entry for every entity targetted by an action, and one entry for every BApp.

In the repository, the **LiveTopologyPaginator** uses this cache to sort the OIDs matching a search, and returns only the next page of results.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41716/