[OM-62911]: XL: Fix tag search resolution in topology processor.
Issue
=====
Groups with tag filters not resolved properly in the topology processor during setting application.

Background
==========
In 7.22.9 I made an optimization to do a "bulk" resolution when resolving tag properties. This greatly sped up tag search resolution in the repository. However, it introduced a problem in the TP.

In the TagPropertyFilter I assumed that there was a common "Tag Index" shared by all entities in the topology. Thus, I selected the first entity's tag index and used it to resolve the tag search. However, in the Topology Processor we do not have a dedicated tag index because all the tags are already in memory per-entity as part of the TopologyEntityDTO. We construct a "singleton" tag index for each entity on demand (not 100% efficient, but makes the interface more compatible). This means that the TagPropertyFilter was only looking at the tags of the first entity in its input stream.

Fix
===
We want to do a hotfix, so minimal impact is ideal. The fix is to collect all the distinct TagIndex objects from all entities fed into the TagPropertyFilter, and run the search against all of them. In the repository this will be the same as before (since there is only one TagIndex). In the topology this will mean running the map search against each single-entity tag index.

In the future we should consider building up an index in the topology processor as well - or reworking the interface. But this should perform well enough, and fixes the issue while changing only one file.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41740/