[OM-62795]: XL Plan: Store plan data in MySQL instead of Arango.
Intent
======
Allow storage of plan data in MySQL instead of Arango.

Background
==========
We want to consolidate our databases - especially for intersight, where each additional database comes with a lot of effort for high availability/backup-restore, etc. Arango in particular has been a problem, and we are not using the "graph" capabilities of Arango much anymore - except in certain plan corner-cases.

Implementation
==============
There are three core places to look for the implementation, all in the Repository Component.
1) **PlanEntityStore** (the interface) and **MySQLPlanEntityStore** (the implementation) for the ingestion and the query implementation.
2) **V1_1__create_plan_tables.sql** which contains the schema definition.
3) **ArangoRepositoryRpcService** and **ArangoSupplyChainRpcService** for the use of **PlanEntityStore** from the gRPC calls that fetch the data.

Pretty much all the other changes are due to the **SupplyChain** refactor of string -> int for entity types - which I did to reduce the storage of these supply chains (as blobs) in the database.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41579/