[OM-62982]: RB:41775 Log instead of tag jaeger trace data for repeated stitching operations
Jaeger tags are key-value pairs suitable for indexing to help with searching for specific spans in a trace. Logs are more informational unstructured data not suitable for indexing. The information being added as tags really should have been a log message. Also add a little bit of trace information for max and average operation length for repeated iterations for a stitching operation.