[OM-63073]: XL API: Fix scenario mapping for MCP plan.
Issue
=====
Unable to create MCP plan scenario.

Background
==========
The "ApiId#needsResolution" method was returning "false" if we had determined that the ID was not <something>. For example, if we had checked "isGroup" earlier, and found it wasn't a group, ApiId#groupInfo would be set to Optional.empty(). This ID still need resolution because we don't know what it's type is - we just know it's not a group. The old code, however, would skip it when resolving whether or not its an entity.

Fix
===
Fix ApiId#needsResolution to check for POSITIVE sets of groupInfo, entityInfo, planInfo, or target.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41802/