[OM-63004]: XL Perf: Minor improvement to CpuScalingFactor post-stitching operation and EntitySettingsApplicator.
Intent
======
Speed up post-stitching, reduce allocation in entity settings application stage.

Background
==========
The CpuScalingFactorPostStitchingOperation was the slowest post-stitching operation in the BofA environment - both for local testing and in the appliances Yue has been monitoring. In that stage we are recursively traversing up from hosts, and then (sometimes) back down, in order to update scaling factors on all relevant entities. With the added ARM entities this can end up being a very large traversal.

The fix is a simple early return - only traverse up/down if we actually modify some commodities on an entity.

In the EntitySettingsApplicator I noticed lots of allocations which trace back to creating a new set of applicators for every entity. This is completely unnecessary because the applicators do not have entity-specific state.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41838