[OM-63209] Prevent setting percentile for commodity if no count data has been recorded for it/Round up rank index during percentile calculation.
Diagnosis
---------
1. Percentile for no recorded counts

It was observed that 0 percentile is set to sold commodity when no count data has been recorded for it. This causes incorrect scale down actions, which are in general guarded by resizable flag being false when the commodity is not monitored. However in this case, due to an issue still under investigation, resizable flag is true and used value is set - both of these fields are being stitched from AppD proxy VM entity.

2. Rank index during percentile calculation

Currently the rank index is calculated using:
int rankIndex = (int)(total * rank / 100);

when there is exactly 1 recorded count - then the rankIndex is evaluated as 0 (1 x 95 / 100 = 0 when rounded to int). The correct solution would be 1, i.e. the 95th percentile is the single recorded utilization count itself. This happens for example when AppD VMem used is stitched to Azure VMem - the resulting 0 percentile causes an incorrect Scale down action.

Summary of change
-----------------
In this review, setting of percentile field in HistoricalValue of Commodity is prevented if no counts have been recorded for the commodity. Further, rounded up the rank index calculation instead of ronuding down.

Review: https://rbcommons.com/s/VMTurbo/r/41924/