[OM-63269] Updated sevone/prometheus-gateway tag in prometheus Helm chart
This was done to pull the fix to DCO-6 and ensure we are not missing
telemetry data due to connection drops.

Deployed updated operator on an XL instance in our lab and let it send
data for 18 hours. Then compared data seen in local Grafana to the data
seen on DataCloud telemetry test project's Grafana dashboards metric by
metric. I no longer see missing data. We also looked at the debug-level
logs with the SevOne engineer and all seems good.