[OM-63276]: RB:41936 Optimize performance of creating StitchingContext
** Problem **
Before we run any stitching operations, we have to first construct a StitchingContext object from the EntityDTOs stored in the EntityStore. This StitchingContext contains a graph that we can stitch on as well as acceleration structures to look up necessary data during stitching. Currently, it takes nearly as long to construct the StitchingContext as it does to run both the pre-stitching and main-stitching operations combined. We should try to optimize the creation of this StitchingContext in order to speed up the time it takes to run the overall topology pipeline.

** Solution **
I made several small optimizations each of which delivered a small win. There was not an obvious place we could immediately deliver a big improvement. The changes are:

1. Don’t get regular built proto and then call toBuilder again for commodity objects. Just get the builder directly. This was done for both commodities bought and sold. Before we were going Builder -> Built -> Builder. Now we just take the original Builder.
2. Optimize the parsing of the accesses key. Before we were calling String#split which compiles a regex when it is called. Now we just call indexOf which does not require compiling a regex.
3. Cache the fetching of lastUpdatedTime and supportsConnectedTo when constructing stitchingDataMap. This helps save a large percentage of the time we were spending when constructing StitchingEntityData.

** Testing Done **
Testing was done on the Merck topology loaded into my local Voltron instance. We probably save about 1/3 of the time we were spending on both the commodities bought and sold part of the StitchingContext creation as a result of these optimizations. Much of the remaining time we are spending here is spent in the simple act of iterating the commodities bought and sold. I tried a number of techniques to optimize this, but a lot of this time seems hard to optimize away. Using an iterator over the Commodity builders instead of a for loop over the built commodities does help some here but probably half the Commodities sold time and a little less of the Commodities bought time is still spent on just the simple act of going through the iterator.

Overall I would guess these optimizations save between 10% of the time we spent constructing the stitching context.

Also all unit tests continue to pass.