[OM-63281]: RB:41951 Add tracing support into market analysis
** Problem **

In our lab instances running the bank of america topology, the market is responsible for nearly half the amount of time it takes between discovery and action execution on the discovered data.

Trying to trace through the timings in the logs is proving challenging and it seems like we are missing a good amount of data. Jaeger traces vastly simplify the process of understanding what is taking up time. It has proven very helpful in our investigations into performance in other components, and the fact that it is missing in analysis has made it harder to conduct an investigation there.

** Solution **

Introduce jaeger tracing into the market analysis (EDE#generateActions).
* Do this by creating a ITracer interface in com.vmturbo.commons. A default implementation of this interface is a no-op so that there will be no overhead when it is unused.
* Pass an optional ITracer instance into EDE#generateActions. In the XL codebase, we pass in an implementation that interacts with jaeger to create distributed traces of a broadcast operation.

** Testing Done **
* Tested locally and see the correct spans for the work being done by analysis.
* Deployed an image to an appliance with the BofA topology loaded. Using the new jaeger data it is now much more apparent why analysis is slow and where the time is going.