[OM-63194]: XL API: Properly retrieve RI actions/action stats.
Issue
=====
BuyRI actions not coming up in Billing Family scope.

Background
==========
When we expand action query scopes we FIRST extract the buy RI entities from the input (e.g. extract the business accounts from the billing family) and then expand "aggregated" entities using the SupplyChainFactory.

Unfortunately, the business accounts in the family count as "aggregated" entities, so they get expanded into their virtual machines, volumes, and so on, and we lose the Business Account ID in the query to the AO.

Fix
===
The fix is to collect the buyRiEntities and add them to the "final" scope without expansion.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41947/