[OM-63345]: XL: Reduce memory footprint of IdentityServiceInMemoryUnderlyingStore.
Issue
=====
On larger topologies the identity service is taking up 3-5GB of heap.

Background
==========
The identity service in the topology processor keeps all the entity descriptors in memory. In addition to the identifying and heuristic properties, each id descriptor has a "query set" which is a concatenated string of all the property descriptors.

An easy optimization is to get rid of this query set. When querying, we can compare the input property descriptors against the identifying and heuristic properties directly. We only run these queries for entities that have volatile properties change to determine whether or not the heuristic threshold is met.

In the future we should strongly consider NOT having all the descriptors in memory, but that would be a more invasive change.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41896/