[OM-63276]: RB:42011 Optimize the convert_to_market phase
** Problem **

In an EKS cluster with the BofA diags loaded we are seeing convert_to_market take 90 seconds. It should be faster, particularly if we want to be able to get to the goal of being able to act on a change in the customer's environment within 10 minutes of it happening. See screenshot below.

** Solution **

Optimize the code.
* The biggest savings were actually produced by changing a debug statement to use lambdas. We were evaluating the toString statement on a commodity in a debug log which was eating up a huge chunk of time.
* Small savings by converting a List<Pair<Float, Float>> into a List<UsedAndPeak> where UsedAndPeak is a simple data structure that wraps up a pair of primitive floats. This also provides some small memory savings.
* Decent savings by taking some very inefficient code that created a new Gson() object in a tight loop and then converted some JSON into a map which was then parsed into doubles. By reusing the same Gson object and being a little more careful about when we use it, it saved a bunch of time. This code used to take about 7 seconds at BofA and now it is taking 0.1s. I turned this into a general utility because it appeared that we were doing the same inefficient thing in other spots in the code so I expect there is some more savings elsewhere too but I did not measure it explicitly.

** Additional notes **

There is additional remaining time we can save around the creation of strings for identifying commodities and assigning them indices. I tried to optimize this away (it would probably save another 3-5% of time, as well as providing memory savings) but it wasn't that easy to do. I would like to take another crack at this later, but this will take more time for only a marginal benefit when there are other more fruitful areas to focus on in the meantime.

** Testing Done **
All testing was done in the EKS cluster for performance testing.

BofA: About a 40% improvement (90s to 50s)
Merck: About 32.5% improvement (40.44s to 27s)
Citi: About a 52% improvement (116s to 55s)