[OM-63380]: Fix performance issue in live topology pipeline
Review RB-42001: https://rbcommons.com/s/VMTurbo/r/42001/

`EntityRetriever` class was obtaining entities from the cached
topology in a manner whose cost grew linearly with the size of the
topology, not the size of the request. In large topologies, locking
could cause `CacheWritingConstructTopologyFromStitchingContextStage`
in the pipeline to be delayed by a few minutes or even as much as 15
minutes while these retrievals kept the cached topology locked.