[OM-63146]: XL: SetAutoSetCommodityCapacityPostStitchingOperation optimizations.
Intent
======
Speed up the SetAutoSetCommodityCapacityPostStitchingOperation, which is the slowest (by far) operation in our BofA appliances.

Background
==========
We are looking up historical capacities of 2 commodities (response time and transaction) of 5 entity types every broadcast. On an appliance with lots of history this becomes VERY slow, because each capacity lookup is a fairly expensive query, and there are around 350k total entities that need it.

Fix
===
The fix is to move the capacity lookup off the critical path, and keep the historical capacities in a lightweight cache. I'm using a:

commodity type -> entity oid -> double

structure because that was about 20% more memory-efficient than the natural entity oid -> commodity -> double structure. This takes up 32MB of memory for 500k entities.

The cache is implemented in the **MaxCapacityCache** utility class under SetAutoSetCommodityCapacityPostStitchingOperation. It is refreshed every 6 hours by default.

Because the cache needs to contain all values, I also changed the RPC method in the history component to accept a list of entity type -> commodity type mappings, and return data for all entities.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/41989/