[OM-62718] Modified ControllableManager to consider maintenance mode.
Now entities that are in maintenance mode are never non-controllable and
so are those that recently exited maintenance mode and weren't touched
by executed actions after that. Entities placed on other entities that
are trying to enter maintenance mode are controllable.

Arguably, entities that move shouldn't themselves be made non-
controllable when that happens. But since their move can effect the
suppliers' quantities sold it's possible that their expenses are
effected and resizes and/or provisions/suspensions are also inaccurate.
Thus I considered making a more general change to not make move targets
non-controllable too risky to be put into a bugfix. We can consider it
for an improvement.

I tested this using Voltron which I connected to lab VC targets. Initial
testing was done on dc3 with host esx24 and final on dc7 with host esx48.
I had reduced broadcast and discovery intervals to 1 minute. The steps
were:

1. Move a VM to the host in question from inside XL UI to make the VM,
   source and destination non-controllable.
2. Request the host be put into maintenance from within vSphere UI.
3. Watch for actions to evacuate the host in XL UI. Take those actions.
4. Once the host is successfully in maintenance mode, request that it
   exits maintenance mode from within vSphere UI.
5. Watch for actions to fill the host in XL UI.

First I reproduced the problem without the fix: in steps 3 and 5 I didn't
see actions until controllable was reset after 30 minutes and I had to
move the VMs manually from within vSphere. That was actually done on an
XL instance, before I setup Voltron.

Then I used the fix and saw the actions in steps 3 and 5 they appeared
in 1-2 minutes but with default broadcast and discovery schedules they
should appear in ~10 minutes.