[OM-63308]: RB:42042 Optimize construction of Ledger in market analysis
** Problem **

In both the first and second rounds of analysis, creating the Ledger (where we calculate revenues and expenses) takes a very long time at large customers. Since we perform two rounds of analysis for realtime, we create the ledger twice every time we do an analysis on a broadcast, which is extremely expensive. Jaeger traces show that on the BofA topology running in an EKS cluster, 21% of the total time spent in analysis is spent creating the Ledger.

** Solution **

There is an O(n^2) operation when creating the ledger. There is a checkArgument call to validate that the economy index on a trader matches its location in the Ledger's list of IncomeStatements. This check is implemented as an O(n) contains call where n is proprtional to the number of traders in the economy. Since we create a TraderIncomeStatement for every Trader, this works out to O(n^2). The solution is to replace the check by just verifying that the item in the economyIndex'th position in the list is the correct income statement, which is a very similar (and probably more correct) form of validation which operates in O(n) time.

This is a slight update on the previous review that slightly changes the checkArgument call based on feedback from Vaptisis.

** Testing Done **

This change just refactors a checkArgument call which had no functional impact since it was not throwing any exceptions before and is not throwing any after the refactor. In Bank of America running in EKS ledger creation used to take about 45 seconds, and was done twice for every round of real-time analysis. Now it takes about ~200 milliseconds.