[OM-63459] ActionInterpreter - Restore old logic for Move actions in interpretAction.
Diagnosis
---------
MoveTOs generated for Volume resize actions (no change in Tier, but change in capacity of some commodities) do not contain any provider changes, but only commodity changes. For such MoveTOs, interpretMoveAction returns an empty Optional and with the current logic results in skipping Action generation for these MoveTOs.

Description of fix
------------------
Reverted to previous logic which first checks whether the MoveTO needs to be translated to Scale action and if so, Scale actions are created for them. Since MoveTOs for Volume resize actions need to be translated to Scale actions, this logic successfully creates Scale actions for them.

Review: https://rbcommons.com/s/VMTurbo/r/42070/