[OM-59834]: Implement the logic for getting parent and siblings for for a business account
Intent
======
Implement the unimplemented logic for getting parent and siblings of an account

Problem
=======
When calling /businessunits/.../relatedBusinessUnits for the siblings and parent, the api still returns the children.

Diagnosis
=========
The logic for getting those has not been implemented and we return children of account regardless of relationship provided in API.

Solution
========
Implement the missing logic.
The logic for getting parent requires iterating all the accounts in the system and check their children. There is no alternative option as there is only one way relationship from parent account to children accounts. That is why this approach was used.