[OM-63444] Exclude undiscovered accounts from cost stats
Through EA billing discovery, costs may be discovered for accounts that are not discovered (i.e. whose Service Principal has not been added as a target). The costs from these undiscovered accounts are stored like any other costs, in case the account is added in the future, but should not be present in cost stats retrieved for the discovered environment.

RB:42127