OM-63633: Fix unable to login as AD group user (rb42168)
Problem:
User is unable to login as AD group user when "enableMultiADGroupSupport" is set to false.

Investigation:
It's regression and seems related to condition check in AD group.

Proposed fixes:
Fix the problematic condition check.

Testing Done
Updated the test to cover the changes
Verified the fixes work on local, 10.10.172.242 and 10.10.170.225 environment.