[OM-63710]: XL TP: Do not retain full topology graph in GroupScopeResolver.
Intent
======
Reduce memory retained by the GroupScopeResolver.

Background
==========
The GroupScopeResolver keeps a reference to the latest TopologyGraph, updated each live topology pipeline. This is used to find an entity's GuestLoad application. The problem is that this prevents the whole TopologyGraph from being garbage collected until the next GroupScopeResolver#setTopologyGraph call.

Fix
===
The simple fix is to process the entities in the graph and keep the entity oid -> guestload oid map in the GroupScopeResolver.

Reviewed At: https://rbcommons.com/s/VMTurbo/r/42270/