[OM-62718] Modified ControllableManager to reduce number of DB queries.
Now `deleteMoveActions` takes a list of OIDs for which to delete the
records and ControllableManager collects the OIDs to make a single query
to the database. That reduces the overhead of communicating with the
DBMS and makes the modification of the database transactional but
introduces the risk of creating a huge query. Fortunately, this list of
OIDs is the intersection of non-controllable and in-maintenance entities
and entities aren't usually put into maintenance in mass, so the risk is
low.

This change was requested by Thiru during review.

Used the Checkerframework's @NonNull annotation instead of the Javax one
as the latter couldn't be applied to type parameters.