[OM-62811]: Plan Orchestrator + Cost - Support VM/Resource Group scopes for OCP Buy RI only
**Background**
The CCA analysis supports analysis on a group of entities. The UI also supports scoping a Buy RI OCP plan to a VM group or resourcce group. This commit updates both plan orchestrator and the cost component to wire the UI's group request into CCA.

**Base scoping of CCA**
The default setting of CCA is to run the analysis for **all relevant workloads** within a scope, regardless of whether the plan is scoped for recommendations for all those entities. This allows the RI coverage calculation stage to correctly take into account worklaods outside the scope of the plan, while the recommendation analysis will only consider workloads within the plan scope. Prior to this commit, the full analysis scope was suboptimial, always including all entities. With this commit, the analysis scope (by default) will scope to all entities relevant to an RI recommendation (i.e. all entities within applicable regions).