[OM-64075]: RB:42419 Reduce market memory consumption by interning protobufs
** Problem **

On the Bank of America topology, the market component consumes quite a lot of memory. One of the main contributors to this is the copy of the topology received from the TopologyProcessor. This copy of the topology remains in-memory during the entire market analysis so that we can use it to construct the projected topology once the market analysis is done.

One way to reduce memory usage is to reduce the amount of memory required to retain these protobufs. One technique for doing this is by sharing a single copy of the same protobuf message across the topology instead of duplicating that message for every time that it is needed. An analysis of the BofA topology indicates that if we eliminate every duplicate message in the topology, we could save about 40% of the memory used (the topology takes up a little over 5 GB of memory, and a little over 2GB could be saved).

** Solution **

Create a utility for interning TopologyEntityDTO protobufs. In order to perform interning efficiently, the utility has been tuned to inspect entities and commodities most likely to be duplicated across the topology while skipping entities and commodities unlikely to contain duplication. I ran profiling across a number of customer topologies to ensure we were getting a reasonable tradeoff of memory savings for the time we were spending for the inspection and interning of protobufs.

Note that if we were able to universally intern all duplicate protobufs, it would be possible to compare protobufs in the same topology with the == operator for reference equality instead of the much more expensive .equals() method. However, because we are only selectively removing duplicates, this is not now possible.

Another, probably superior approach to attempting to intern duplicate protobuf messags would be to do so during the actual deserialization process when we turn the raw bytes coming from Kafka back into in-memory protobufs. However, a preliminary investigation showed that this would be quite challenging to do. If necessary, this is something we could look into in the future beause it would permit us to run the interning with less overhead.

Right now we only intern TopologyEntityDTO protobufs in the market when we receive them from the TopologyProcessor. However, a similar technique could probably be applied in other places. For example, TraderTO's in the market. Or if we keep entire TopologyEntityDTO's in the cost component, those could be interned as well. SoldCommodity in RepoGraphEntity in repository might also benefit. There are probably other additional places too that I haven't looked into yet. The Duplicates utilities submitted as part of this review could be used to identify how much any of these areas might be able to benefit from de-duplication.

** Details **

In order to achieve these results, I built some tooling which has general utility and I have included in the review because there are quite a few other places where we could eliminate object duplication in memory through some sort of interning to reduce memory consumption. These tools are helpful in order to identify these places and tune any custom interning logic to be as effective as possible on real customer topologies. The tools are:

1. A set of timing utilities useful for profiling how long various operations take.
2. A set of Duplicates utilities that can be used to identify duplicate objects in-memory. There is a specialization specific to identifying duplicate protobufs, but it can actually be used on any type of object because it uses general Java reflection to traverse the graph of objects in-memory in order to find duplicates.

I have also added two additional jaeger spans - one to capture the amount of time it takes to receive the topology in the market. You can see in the screenshot attached to this review that this time matches the time taken by the TP to send the topology. Since we can't recive the topology faster than it is sent, this shows that the interning is not adding any additional time to the end-to-end broadcast cycle. Further, I added a span to capture building analysis results after the first real-time market subcycle. This is taking quite a bit of time at Bank of America and may be a target for future optimization.

** Performance Notes **

Interning every duplicate message in the entire topology using reflection (code not included in this review, but you can see code for it in the hackathon/protobuf-interning branch in gitlab) saves 40% of memory usage at Bank of America, but takes too long. The code included here only achieves a 20% savings, but runs in 5s instead of 70-90s. Further, because the time spent to run interning occurs while receiving the topology from Kafka, and Kafka takes 30+ seconds to send the topology (we are processing the results in parallel), it does not actually add any additional time at all to the end-to-end time to get market analysis results.

Some specific timings:
At BofA: Topology interning results: Took PT4.305482831S [874,761 total; 589,601 skipped, 1 not interned, 285,159 partially interned (32.60%)]
At Citi: Topology interning results: Took PT7.436015482S [638,113 total; 206,917 skipped, 4 not interned, 431,192 partially interned (67.57%)]
At Humana: Topology interning results: Took PT1.823166495S [353,510 total; 206,446 skipped, 2 not interned, 147,062 partially interned (41.60%)]

** Testing Done **

* Tested on a number of different customer topologies to ensure that we get a reasonable tradeoff of memory savings for the time we spend. At Bank of America, we save 1.1 GB by spending about 5 seconds, with almost all of those 5 seconds occuring while we receive the customer topology so it does not result in any net additional waiting for market analysis results. See attachments on the review for additional details about how much memory is saved per message-type by this interning at Bank of America.
* Unit tests for all new utilities.