[OM-63348]: RB:42320 Reuse action classifier across market subcycles
** Problem **

Creating a new action classifier involves cloning the entire economy, including every trader, basket, market, shopping list, etc. This is an expensive operation that we currently perform twice for the realtime market (once for each market subcycle). This slows down the time for running market analysis.

** Solution **

The action classifier is used to determine which actions can be executed and which can't. The only actions that actually modify the simulation economy cloned in the action classifier are placement actions, and the only placement actions we use are from the first market analysis subcycle. Therefore, it should be safe to reuse the action classifier across subcycles.

** Testing Done **

* Tested on the BofA topology and confirmed that analysis is faster because the classifier only has to be created once instead of twice.
* Hongyue used a test topology and the market diagnostics, confirming that the actions were almost entirely the same whether we created a new action classifier for each subcycle or we reused the classifier from the first subcycle again in the second subcycle.
* All unit tests continue to pass.