[OM-63578]: Fix the issue that the action types sent to actionscript probe are inconsistent with the action types seen in API
Problem
=======
API reports VM resize actions as "RESIZE", while SDK probe that sends action data to the automation policy action script reports them as "RIGHT_SIZE", causing programmatic lookups to fail based on the reported actionType from the SDK.

Diagnosis
=========
We have different action types when communicating with the probe. This was not visible as we mainly implement the probes. However, we now send the actions to the actionscript probe and becomes visible.

Solution
========
The ideal solution should be translation of the action type in the mediation to be consistent with the API when we send the action to the probe. However, that requires significant testing as it requires testing all the probes affected. Therefore, the chosen solution is that doing the translation in mediation. However, it will only translate to the way that is consistent with API when the probe indicates that it supports this type of translation. For now, only actionscript probe supports this type of translation. As times passes, we can change and test more probes to be consistent with API.

Detailed changes:
OpsMgr Repo:
- I added a new new message and property named ``ActionCapabilities`` to ``ProbeInfo`` protobuf message. This message currently has a boolean flag which indicates in the probe supports new action types consistent with API. More flags can be added in future if we have similar situations in future.
- I changed ``IActionExecutor`` interface which is implemented by the probes that execute actions to have new method which returns ``ActionCapabilities`` object. There is a default implement for this method for this interface which returns default instance of ``ActionCapabilities`` .
- I made a change to populate the ProbeInfo with the value returned from the call to the corresponding method in ``IActionExecutor``.

XL Repo:
- I changed the actionscript probe to set the flag for this probe.
- I introduced a method called ``convertToApiConsistentActionType`` in ``ActionConversions`` which converts the internal action type to a the action type consistent with what is seen in API.
- I changed the logic in method ``getSDKActionType`` in ``AbstractActionExecutionContext`` which checks the probe information for the target executing this action to see if it supports API consistent action types. If so, it will call the method newly introduced method in ``ActionConversion``.