[OM-63933] Insert and update records related to Unattached Volumes in the Volume Attachment History table.
Requirements
------------
For Unattached Volumes, we need to track the first time it was discovered. This will enable us to calculate a lower bound on the number of days that the Volumes has remained unattached.

In order to do this, Unattached Volumes much be stored in the Volume Attachment History table with a placeholder VM OID. The Last Attached Date column must be inserted the first time and never updated while processing subsequent topologies.

Summary of changes
------------------
Introduced a new DbInserter which can exclude specific fields from being updated.
Added logic in VolumeAttachmentHistoryWriter that keeps track of Volume not seen attached to any VM. In the finish stage of the Writer, these Volumes are deemed to be Unattahced and written into the Volume Attachment History table with a placeholder VM oid.

Review: https://rbcommons.com/s/VMTurbo/r/42362/