[OM-63859]: Fix the issue that execution schedule for the default policy goes away after a restart.
Intent
======
Prevent default setting values from being deleted initially.

Problem
=======
The user set execution schedule for vcpu resize up in the default policy. After the pod restart the execution schedules goes away.

Diagnosis
=========
We have a logic in ``DefaultSettingPolicyCreator#mergePolicies`` which merges the default policies coming from the code and policies from DB. The This logic removes the setting from db that are no longer in code because that means that setting has been retired. However, the list of setting that come to this method is only those default setting that don't have non-empty values. Therefore, it will not included execution schedule as it empty by default. Therefore, this logic assumes the that execution schedule is no longer a valid setting and gets removed.

Solution
========
I introduced a new method called ``createEntityTypeToSettingSpecNamesMap`` which gives a map of entity types to setting specs name which contains all settings applicable to that entity type and will use that map when trying to find those setting that are no longer relevant.