[OM-63978]: Fix the issue that actions stats don't come up with cluster
Problem
========
The stats information does not get populated in both Citi and BofA diags.

Diagnosis
=========
The issue happens because there a place in the code where we try to merge all the groups in scope of query to a map. The issue is that the collector in the map does not have merge logic. In the case that two of those groups have same member, that collector throws an exception.

Solution
========
Simply add a merge policy which goes with the first value as in case of collision the values are exactly the same.